requirements.txt
annotated-types==0.7.0
antiorm==1.2.1
anyio==4.10.0
certifi==2025.8.3
db==0.1.1
db-sqlite3==0.0.1
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
numpy
openai==1.107.0
pandas
pip==25.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
python-dotenv==1.0.1
SQLAlchemy==2.0.36
langchain==0.3.7
langchain-openai==0.2.6
